--- 0027/library0027/src/main/java/ru/otus/megration/service/GenreService.java
@@ -1,12 +1,0 @@
-package ru.otus.megration.service;
-
-import org.springframework.stereotype.Service;
-import ru.otus.domain.Genre;
-
-@Service
-public class GenreService {
-    public ru.otus.megration.entity.Genre convert(Genre genre) {
-        return new ru.otus.megration.entity.Genre(genre.getName());
-    }
-
-}
