+++ 0027/library0027/src/main/java/ru/otus/migration/BatchCommands.java
@@ -1,4 +1,4 @@
-package ru.otus.megration;
+package ru.otus.migration;
 
 
 import lombok.RequiredArgsConstructor;
