+++ 0027/library0027/src/main/java/ru/otus/migration/service/AuthorService.java
@@ -1,0 +1,12 @@
+package ru.otus.migration.service;
+
+import org.springframework.stereotype.Service;
+import ru.otus.domain.Author;
+
+@Service
+public class AuthorService {
+    public ru.otus.migration.entity.Author convert(Author author) {
+        return new ru.otus.migration.entity.Author(author.getName());
+    }
+
+}
